Fix PC/IP byte-width bug capping every program at 256 bytes
Root cause: VM.cs declared PC and IP as `byte` (max 255), while
instructions are a fixed 6 bytes with no size cap, RAM defaults to
1 MB (Globals.DefaultRamSize), and the X/Y addressing registers are
already Int32. Once code crossed the 255-byte boundary mid-instruction,
unchecked byte arithmetic silently wrapped PC/IP back to 0, resuming
execution at a misaligned offset — silent output truncation or an
outright "Unsupported instruction" crash on the garbage-decoded opcode,
depending on where the wrap landed. Every JMP/CALL/RET/JMT/JMF/CLT/CLF
target was additionally hard-truncated via explicit (byte) casts in
op.cs, independent of PC's own field width.

Fix: widen PC and IP from byte to int (VM.cs), matching X/Y's existing
width; drop SetRegister's byte-truncation of PC/IP writes and the
GetRegister casts (VMextended.cs), and remove every (byte) cast on a
jump/call/return target in op.cs. DebugForm.cs's memory view needed
one matching widen (byte ip -> int ip) to keep building.

TDD: added LargeProgram_ExceedingFormerByteAddressCeiling_
ExecutesWithoutCorruption (600-byte program, well past the former
255-byte ceiling) to AIL-Tests. Watched it fail two different ways
first — reproducing the exact "Unsupported instruction: opcode 0x16"
crash from manual testing, and then (after the type fix, before a
mistake in the test's own program was caught) a silent no-op — before
it went green. Removed the now-inaccurate comment on
Calculator_AllOperations_PrintsAllResults that documented the wrap as
expected behavior and told programs to stay under 256 bytes.

Also fixed Globals.DebugMode's default, which was hardcoded `true`
against its own doc comment ("Set to false for normal execution"),
so every fresh KEI 0x01 call emitted "KEI 0x01: ..." trace noise into
stdout by default. All existing tests already set DebugMode = false
explicitly, so this had zero effect on the suite — only on real usage
via AIL-Runtime.

Docs: wiki/Spec-Architecture.md and wiki/Specification.md both listed
PC/IP as "8-bit" in the register table; corrected to 32-bit to match
the fix. Note: wiki/Specification.md fully duplicates the content of
the three split spec pages inline below its table of contents (460
lines vs. the split pages' combined ~440) rather than just linking to
them — pre-existing from the doc-split commit, not touched here beyond
the two corrected facts, but worth collapsing to just the TOC/links to
stop this kind of drift. Also pre-existing and unrelated to this fix:
wiki/Spec-Architecture.md §1 describes a 16-bit/64 KB address space
with an Interrupt Vector Table at 0x0000-0x01FD; no IVT is implemented
anywhere in source, and RAM has defaulted to 1 MB since commit 9e73a65.
Left alone — rewriting the memory-map section is a spec/architecture
decision, not part of this bugfix.

Verified: full solution build (0 warnings/errors), full test suite
(43/43), and the real AIL-Runtime binary against examples/calculator.ail
end-to-end — now prints all four operations cleanly with no debug
noise, where it previously truncated silently after the second result.

+++ source/AIL-Studio/DebugForm.cs
@@ -264,7 +264,7 @@ namespace AIL_Studio
             _memBox.Clear();
 
             int codeEnd = _vm.ram.RAMLimit;
-            byte ip     = _vm.IP;
+            int ip      = _vm.IP;
 
             // Header
             int hs = _memBox.TextLength;
